+++ src/main/java/be/baur/sdt/xpath/SDAXPath.java
@@ -10,11 +10,14 @@ import be.baur.sdt.xpath.function.CompareStringFunction;
 import be.baur.sdt.xpath.function.CurrentDateTimeFunction;
 import be.baur.sdt.xpath.function.DateTimeFunction;
 import be.baur.sdt.xpath.function.DocumentNodeFunction;
+import be.baur.sdt.xpath.function.FormatDateTimeFunction;
 import be.baur.sdt.xpath.function.LeftFunction;
+import be.baur.sdt.xpath.function.MillisToDateTimeFunction;
 import be.baur.sdt.xpath.function.ParseSDAFunction;
 import be.baur.sdt.xpath.function.RenderSDAFunction;
 import be.baur.sdt.xpath.function.RightFunction;
 import be.baur.sdt.xpath.function.StringJoinFunction;
+import be.baur.sdt.xpath.function.TimestampFunction;
 import be.baur.sdt.xpath.function.TokenizeFunction;
 
 /**
@@ -44,10 +47,13 @@ public class SDAXPath extends BaseXPath {
 		ctx.registerFunction(SDT.W3CFUNCTIONS_NS_URI, "current-dateTime", new CurrentDateTimeFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "document-node", new DocumentNodeFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "dateTime", new DateTimeFunction());
+		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "format-dateTime", new FormatDateTimeFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "left", new LeftFunction());
+		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "millis-to-dateTime", new MillisToDateTimeFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "parse-sda", new ParseSDAFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "right", new RightFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "render-sda", new RenderSDAFunction());
+		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "timestamp", new TimestampFunction());
 		ctx.registerFunction(SDT.FUNCTIONS_NS_URI, "tokenize", new TokenizeFunction());
 		ctx.registerFunction(SDT.W3CFUNCTIONS_NS_URI, "string-join", new StringJoinFunction());
 	}